requirements.txt
google_api_python_client==1.12.8
google_auth_oauthlib==0.4.2
instaloader==4.5.5
instalooter==2.4.4
moviepy==1.0.3
protobuf==3.14.0
python_dateutil==2.8.1
schedule==0.6.0
pytrends==4.9.0
